+++ src/input.h
@@ -9,7 +9,7 @@ class Input
     GLFWwindow* m_glwindow;
 
     static bool m_rightMouseDown, m_leftMouseDown;
-    static float m_scrollOffset, m_relScroll, m_cursorX, m_cursorY, m_relCursorX, m_relCursorY;
+    static double m_scrollOffset, m_relScroll, m_cursorX, m_cursorY, m_relCursorX, m_relCursorY;
 
     static void key_callback(GLFWwindow* window, int key, int scancode, int action, int mods);
     static void mouse_button_callback(GLFWwindow* window, int button, int action, int mods);
@@ -21,12 +21,12 @@ public:
     void poll(Camera& cam);
     static bool rightMouseDown();
     static bool leftMouseDown();
-    static float scrollOffset();
-    static float relScroll();
-    static float cursorX();
-    static float cursorY();
-    static float relCursorX();
-    static float relCursorY();
+    static double scrollOffset();
+    static double relScroll();
+    static double cursorX();
+    static double cursorY();
+    static double relCursorX();
+    static double relCursorY();
     bool getKey(int key);
     const int* begin();
     const int* end();
